+++ system/src/Grav/Framework/Flex/Pages/Traits/PageContentTrait.php
@@ -673,7 +673,7 @@ trait PageContentTrait
         $process_twig = $this->shouldProcess('twig') || $this->isModule();
         $cache_enable = $this->getNestedProperty('header.cache_enable') ?? $config->get('system.cache.enabled', true);
 
-        $twig_first = $this->getNestedProperty('header.twig_first') ?? $config->get('system.pages.twig_first', true);
+        $twig_first = $this->getNestedProperty('header.twig_first') ?? $config->get('system.pages.twig_first', false);
         $never_cache_twig = $this->getNestedProperty('header.never_cache_twig') ?? $config->get('system.pages.never_cache_twig', false);
 
         $cached = null;