+++ system/src/Grav/Common/Flex/Types/Pages/Traits/PageLegacyTrait.php
@@ -192,9 +192,6 @@ trait PageLegacyTrait
             throw new InvalidArgumentException('Argument should be either header variable name or array of parameters');
         }
 
-        if (!$pagination) {
-            $params['pagination'] = false;
-        }
         $context = [
             'pagination' => $pagination,
             'self' => $this