+++ system/defines.php
@@ -8,9 +8,9 @@
 
 // Some standard defines
 define('GRAV', true);
-define('GRAV_VERSION', '1.7.0-rc.20');
+define('GRAV_VERSION', '1.7.0');
 define('GRAV_SCHEMA', '1.7.0_2020-11-20_1');
-define('GRAV_TESTING', true);
+define('GRAV_TESTING', false);
 
 // PHP minimum requirement
 if (!defined('GRAV_PHP_MIN')) {