+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     * Fixed Flex sorting issues [grav-plugin-flex-objects#92](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/92)
     * Fixed Clockwork missing dumped arrays and objects
     * Fixed fatal error in PHP 8 when trying to access root page
+    * Fixed Array->String conversion error when `languages:translations: false` [admin#1896](https://github.com/getgrav/grav-plugin-admin/issues/1896)
 
 # v1.7.0-rc.20
 ## 12/15/2020