+++ CHANGELOG.md
@@ -25,7 +25,8 @@
     * Fixed fatal error in PHP 8 when trying to access root page
     * Fixed Array->String conversion error when `languages:translations: false` [admin#1896](https://github.com/getgrav/grav-plugin-admin/issues/1896)
     * Fixed `Inflector` methods when translation is missing `GRAV.INFLECTOR_*` translations
-    * Fixed `PageObject::getRoute()` with pages which have no route
+    * Fixed exception when changing parent of new page [grav-plugin-admin#2018](https://github.com/getgrav/grav-plugin-admin/issues/2018)
+    * Fixed ordering when changing parent of new page [grav-plugin-admin#2018](https://github.com/getgrav/grav-plugin-admin/issues/2018)
 
 # v1.7.0-rc.20
 ## 12/15/2020